# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cube-back",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/liensRessourceUserArchive/archivesFromUser/{id_user}": {
            "get": {
                "tags": [
                    "LienRessourceUserArchive"
                ],
                "summary": "Récupérer tous les archives d'un utilisateur",
                "operationId": "8e1d063707454f46c07e9baec441676c",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items to return per page (1-100)",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword for searching items",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort items by",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/liensRessourceUserArchive/{id_res}/{id_user}": {
            "get": {
                "tags": [
                    "LienRessourceUserArchive"
                ],
                "summary": "Récupérer une lien d'archivage entre la ressource et l'utilisateur",
                "operationId": "233e431dccd556546b89c2a699ece432",
                "parameters": [
                    {
                        "name": "id_res",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true
                    },
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "LienRessourceUserArchive"
                ],
                "summary": "Supprimer une ressource des archives",
                "operationId": "8712d8ab643737a3394fdca9879fdf54",
                "parameters": [
                    {
                        "name": "id_res",
                        "in": "path",
                        "description": "ID of the ressource",
                        "required": true
                    },
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item deleted successfully"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/liensRessourceUserArchive": {
            "post": {
                "tags": [
                    "LienRessourceUserArchive"
                ],
                "summary": "Mettre en archive une ressource",
                "operationId": "946fc718a4fdf8c625857bd3874afcef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_res",
                                    "id_user"
                                ],
                                "properties": {
                                    "id_res": {
                                        "type": "integer"
                                    },
                                    "id_user": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}